+++ src/ByteDev.Http/HttpContentHeadersExtensions.cs
@@ -15,6 +15,7 @@ namespace ByteDev.Http
         /// <param name="source">Headers collection to perform the operation on.</param>
         /// <param name="name">The header to add or update in the collection.</param>
         /// <param name="value">The content of the header.</param>
+        /// <exception cref="T:System.ArgumentNullException"><paramref name="source" /> is null.</exception>
         public static void AddOrUpdate(this HttpContentHeaders source, string name, string value)
         {
             if (source == null)