+++ tests/ByteDev.Http.UnitTests/FormUrlEncoded/Serialization/TestDummyBuiltInRefTypes.cs
@@ -1,9 +1,0 @@
-﻿namespace ByteDev.Http.UnitTests.FormUrlEncoded.Serialization
-{
-    internal class TestDummyBuiltInRefTypes
-    {
-        public object Obj { get; set; }
-
-        public string String { get; set; }
-    }
-}