+++ tests/ByteDev.Http.UnitTests/HttpContentExtensionsTests.cs
@@ -2,8 +2,6 @@
 using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
-using ByteDev.Http.FormUrlEncoded.Serialization;
-using ByteDev.Http.Xml;
 using NUnit.Framework;
 
 namespace ByteDev.Http.UnitTests
@@ -53,29 +51,6 @@ namespace ByteDev.Http.UnitTests
                 var sut = new StringContent(xml);
 
                 var result = await sut.ReadAsXmlAsync<TestDummy>();
-
-                Assert.That(result.Name, Is.EqualTo(dummy.Name));
-            }
-        }
-
-        [TestFixture]
-        public class ReadAsFormUrlEncodedAsync : HttpContentExtensionsTests
-        {
-            [Test]
-            public void WhenContentIsNull_ThenThrowException()
-            {
-                Assert.ThrowsAsync<ArgumentNullException>(() => HttpContentExtensions.ReadAsFormUrlEncodedAsync<TestDummy>(null));
-            }
-
-            [Test]
-            public async Task WhenContentIsFormUrlEncoded_ThenReturnObj()
-            {
-                var dummy = new TestDummy {Name = "John"};
-                var form = FormUrlEncodedSerializer.Serialize(dummy);
-
-                var sut = new StringContent(form);
-
-                var result = await sut.ReadAsFormUrlEncodedAsync<TestDummy>();
 
                 Assert.That(result.Name, Is.EqualTo(dummy.Name));
             }